# apps/api-java/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "ChatSpace API",
    "description" : "Slack風チャットアプリ ChatSpace のREST API。詳細はdocs/要件定義書.mdを参照。",
    "version" : "v1"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
  "paths" : {
...
    "/workspaces/{workspaceId}/channels/{channelId}/messages/{messageId}/reactions" : {
      "post" : {
        "tags" : [ "message-controller" ],
        "operationId" : "toggleReaction",
        "parameters" : [ {
          "name" : "workspaceId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "channelId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "messageId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ToggleReactionRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MessageResponse"
                }
              }
            }
          }
        }
      }
    },
    "/workspaces/{workspaceId}/dms/{dmId}/messages/{messageId}/reactions" : {
      "post" : {
        "tags" : [ "message-controller" ],
        "operationId" : "toggleReaction_1",
        "parameters" : [ {
          "name" : "workspaceId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "dmId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "messageId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ToggleReactionRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MessageResponse"
                }
              }
            }
          }
        }
      }
    },
    "/workspaces/{workspaceId}/channels/{channelId}/messages" : {
      "get" : {
        "tags" : [ "message-controller" ],
        "operationId" : "list_3",
        "parameters" : [ {
          "name" : "workspaceId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "channelId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "cursorCreatedAt",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date-time"
          }
        }, {
          "name" : "cursorId",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MessageListResponse"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "message-controller" ],
        "operationId" : "create_2",
        "parameters" : [ {
          "name" : "workspaceId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "channelId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateMessageRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MessageResponse"
                }
              }
            }
          }
        }
      }
    },
    "/workspaces/{workspaceId}/dms/{dmId}/messages" : {
      "get" : {
        "tags" : [ "message-controller" ],
        "operationId" : "list_4",
        "parameters" : [ {
          "name" : "workspaceId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "dmId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "cursorCreatedAt",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date-time"
          }
        }, {
          "name" : "cursorId",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MessageListResponse"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "message-controller" ],
        "operationId" : "create_3",
        "parameters" : [ {
          "name" : "workspaceId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "dmId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateMessageRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MessageResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/workspaces/{workspaceId}/dms/{dmId}/messages/{messageId}" : {
      "delete" : {
        "tags" : [ "message-controller" ],
        "operationId" : "delete",
        "parameters" : [ {
          "name" : "workspaceId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "dmId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "messageId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      },
      "patch" : {
        "tags" : [ "message-controller" ],
        "operationId" : "edit",
        "parameters" : [ {
          "name" : "workspaceId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "dmId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "messageId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/EditMessageRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MessageResponse"
                }
              }
            }
          }
        }
      }
    },
    "/workspaces/{workspaceId}/channels/{channelId}/messages/{messageId}" : {
      "delete" : {
        "tags" : [ "message-controller" ],
        "operationId" : "delete_1",
        "parameters" : [ {
          "name" : "workspaceId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "channelId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "messageId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      },
      "patch" : {
        "tags" : [ "message-controller" ],
        "operationId" : "edit_1",
        "parameters" : [ {
          "name" : "workspaceId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "channelId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "messageId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/EditMessageRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MessageResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/workspaces/{workspaceId}/dms/{dmId}/messages/{messageId}/replies" : {
      "get" : {
        "tags" : [ "message-controller" ],
        "operationId" : "replies",
        "parameters" : [ {
          "name" : "workspaceId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "dmId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "messageId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "cursorCreatedAt",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date-time"
          }
        }, {
          "name" : "cursorId",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MessageListResponse"
                }
              }
            }
          }
        }
      }
    },
    "/workspaces/{workspaceId}/channels/{channelId}/messages/{messageId}/replies" : {
      "get" : {
        "tags" : [ "message-controller" ],
        "operationId" : "replies_1",
        "parameters" : [ {
          "name" : "workspaceId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "channelId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "messageId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "cursorCreatedAt",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date-time"
          }
        }, {
          "name" : "cursorId",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MessageListResponse"
                }
              }
            }
          }
        }
      }
    },
    "/workspaces/{workspaceId}/dms/{dmId}/messages/{messageId}/context" : {
      "get" : {
        "tags" : [ "message-controller" ],
        "operationId" : "context",
        "parameters" : [ {
          "name" : "workspaceId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "dmId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "messageId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MessageContextResponse"
                }
              }
            }
          }
        }
      }
    },
    "/workspaces/{workspaceId}/channels/{channelId}/messages/{messageId}/context" : {
      "get" : {
        "tags" : [ "message-controller" ],
        "operationId" : "context_1",
        "parameters" : [ {
          "name" : "workspaceId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "channelId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "messageId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MessageContextResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "ToggleReactionRequest" : {
        "type" : "object",
        "properties" : {
          "emoji" : {
            "type" : "string",
            "maxLength" : 8,
            "minLength" : 1
          }
        },
        "required" : [ "emoji" ]
      },
      "MessageResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "channelId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "dmId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "parentId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "authorId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "body" : {
            "type" : "string"
          },
          "deleted" : {
            "type" : "boolean"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "updatedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "editedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "replyCount" : {
            "type" : "integer",
            "format" : "int64"
          },
          "reactions" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ReactionSummary"
            }
          }
        }
      },
      "ReactionSummary" : {
        "type" : "object",
        "properties" : {
          "emoji" : {
            "type" : "string"
          },
          "count" : {
            "type" : "integer",
            "format" : "int64"
          },
          "reactedByMe" : {
            "type" : "boolean"
          },
          "userIds" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "format" : "uuid"
            }
          }
        }
      },
      "CreateMessageRequest" : {
        "type" : "object",
        "properties" : {
          "body" : {
            "type" : "string",
            "maxLength" : 4000,
            "minLength" : 1
          },
          "parentId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "attachmentIds" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "format" : "uuid"
            },
            "maxItems" : 10,
            "minItems" : 0
          }
        },
        "required" : [ "body" ]
      },
...
      "EditMessageRequest" : {
        "type" : "object",
        "properties" : {
          "body" : {
            "type" : "string",
            "maxLength" : 4000,
            "minLength" : 1
          }
        },
        "required" : [ "body" ]
      },
      "Cursor" : {
        "type" : "object",
        "properties" : {
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "id" : {
            "type" : "string",
            "format" : "uuid"
          }
        }
      },
...
      "MessageListResponse" : {
        "type" : "object",
        "properties" : {
          "messages" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/MessageResponse"
            }
          },
          "nextCursor" : {
            "$ref" : "#/components/schemas/Cursor"
          }
        }
      },
      "MessageContextResponse" : {
        "type" : "object",
        "properties" : {
          "messages" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/MessageResponse"
            }
          },
          "targetMessageId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "replyMessageId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "olderCursor" : {
            "$ref" : "#/components/schemas/Cursor"
          },
          "newerCursor" : {
            "$ref" : "#/components/schemas/Cursor"
          },
          "hasOlder" : {
            "type" : "boolean"
          },
          "hasNewer" : {
            "type" : "boolean"
          }
        }
      },
...
    }
  }
}